chore: move constants to root
