--- src/widgets/BhandaraCheckin/SectionMain/startCheckin/constants.ts
@@ -1,1 +1,0 @@
-export const onFileText = "**on file**";
